requirements.txt
asgiref==3.5.0
asyncpg==0.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SQLAlchemy==1.4.35
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
